backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
rq==1.16.0
boto3==1.34.34
pymupdf==1.23.21
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.3
python-dotenv==1.0.0
email-validator>=2.0.
openai>=1.0.0
bcrypt==4.0.1

